ajout de end point config avec son controleur
